requirements.txt
# Application
Django==3.2.2
gunicorn==20.1.0
dj-database-url==0.5.0
psycopg2==2.8.6
requests==2.25.1
djangorestframework==3.12.4
pydantic==1.8.1

# Testing
pytest==6.2.4
pytest-django==4.2.0
factory-boy==3.2.0

# Development
ipdb==0.13.7
